Add static cast to remove warning about loss of precision
r=dumi
Review URL: http://codereview.chromium.org/4552001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@65165 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0c34137ded36b4519c8f793386e7115a55e209bf

+++ chrome/browser/jankometer.cc
@@ -144,8 +144,8 @@ JankObserverHelper::JankObserverHelper(
       1, 3600000, 50, base::Histogram::kUmaTargetedHistogramFlag);
   if (discard_count_ > 0) {
     // Select a vaguely random sample-start-point.
-    events_till_measurement_ =
-        (TimeTicks::Now() - TimeTicks()).InSeconds() % (discard_count_ + 1);
+    events_till_measurement_ = static_cast<int>(
+        (TimeTicks::Now() - TimeTicks()).InSeconds() % (discard_count_ + 1));
   }
 }
